# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy ESPN Service API
  description: "Integrates with the ESPN Fantasy v3 API: reads a user's fantasy hockey\
    \ league settings (and stores the espn_s2 / SWID cookies needed for private leagues)\
    \ for the BFF."
  version: v1
servers:
- url: /
...
paths:
...
  /api/v1/espn/players/sync:
    post:
      tags:
      - ESPN Players
      summary: Refresh the cached player read model from ESPN
      operationId: sync
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSyncResponse"
        "502":
          description: ESPN was unreachable or returned nothing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSyncResponse"
...
  /api/v1/espn/players/skaters:
    get:
      tags:
      - ESPN Players
      summary: All skaters with eligible positions and a season's stats
      operationId: skaters
      parameters:
      - name: season
        in: query
        description: "Season start year — 2025 is the 2025-26 season. Required: the\
          \ season being collected and the season a caller wants to show are deliberately\
          \ different for most of the year, so there is no sensible default. A player\
          \ with no line for it comes back with no stats rather than being left out."
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Skaters returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkaterResponse"
  /api/v1/espn/players/goalies:
    get:
      tags:
      - ESPN Players
      summary: All goalies with eligible positions and a season's stats
      operationId: goalies
      parameters:
      - name: season
        in: query
        description: "Season start year — 2025 is the 2025-26 season. Required: the\
          \ season being collected and the season a caller wants to show are deliberately\
          \ different for most of the year, so there is no sensible default. A player\
          \ with no line for it comes back with no stats rather than being left out."
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Goalies returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalieResponse"
...
components:
  schemas:
...
    PlayerSyncResponse:
      type: object
      properties:
        players:
          type: integer
          format: int32
          description: Stat lines stored
        syncedAt:
          type: string
          format: date-time
      required:
      - players
      - syncedAt
...
    SkaterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ESPN's player id
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        eligiblePositions:
          type: array
          description: "ESPN fantasy eligible positions (e.g. [\"C\",\"LW\"])"
          items:
            type: string
        sweaterNumber:
          type: integer
          format: int32
        teamAbbrev:
          type: string
        headshot:
          type: string
          description: "ESPN image CDN URL for the player's headshot, already sized\
            \ for the player table"
        gamesPlayed:
          type: integer
          format: int32
        goals:
          type: integer
          format: int32
        assists:
          type: integer
          format: int32
        points:
          type: integer
          format: int32
        plusMinus:
          type: integer
          format: int32
        pim:
          type: integer
          format: int32
        powerPlayGoals:
          type: integer
          format: int32
        powerPlayPoints:
          type: integer
          format: int32
        shorthandedGoals:
          type: integer
          format: int32
        shorthandedPoints:
          type: integer
          format: int32
        gameWinningGoals:
          type: integer
          format: int32
        shots:
          type: integer
          format: int32
        shootingPctg:
          type: number
          format: double
        avgToi:
          type: string
        faceoffWinningPctg:
          type: number
          format: double
        hits:
          type: integer
          format: int32
        blockedShots:
          type: integer
          format: int32
        totalFaceoffWins:
          type: integer
          format: int32
        totalFaceoffLosses:
          type: integer
          format: int32
        hatTricks:
          type: integer
          format: int32
          description: Stats Yahoo does not report at all
        shifts:
          type: integer
          format: int32
        timeOnIce:
          type: integer
          format: int32
          description: "Total time on ice for the season, in seconds"
      required:
      - eligiblePositions
      - firstName
      - id
      - lastName
      - position
    GoalieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ESPN's player id
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        eligiblePositions:
          type: array
          description: "ESPN fantasy eligible positions (e.g. [\"G\"])"
          items:
            type: string
        sweaterNumber:
          type: integer
          format: int32
        teamAbbrev:
          type: string
        headshot:
          type: string
          description: "ESPN image CDN URL for the player's headshot, already sized\
            \ for the player table"
        gamesPlayed:
          type: integer
          format: int32
        gamesStarted:
          type: integer
          format: int32
        wins:
          type: integer
          format: int32
        losses:
          type: integer
          format: int32
        shutouts:
          type: integer
          format: int32
        shotsAgainst:
          type: integer
          format: int32
        saves:
          type: integer
          format: int32
        goalsAgainst:
          type: integer
          format: int32
        goalsAgainstAvg:
          type: number
          format: double
        savePctg:
          type: number
          format: double
        overtimeLosses:
          type: integer
          format: int32
          description: Stats Yahoo does not report at all
        timeOnIce:
          type: integer
          format: int32
          description: "Total time on ice for the season, in seconds"
      required:
      - eligiblePositions
      - firstName
      - id
      - lastName
      - position
...